Add basic cask detection
First round of attempts at parsing the Stardew Valley save file and
detecting the casks in the basement.  At the moment this is pretty
hacked and uses the presence of <agingRate> inside of the generic
<Object> in order to decide if the object is a cask.  This then uses
linq to xml to access data about the casks like item left and days to
maturity.  Though it seems buggy at the moment.