Refactor cask detection by properly using xsi:type
Rather than trying to see if certain tags exist under the cask object
(brittle) we can use the xsi namespace in the XML file to directly check
the xsi:type of each object and actually look for Casks.  Unlikely that
the underlying type of the object will not be Cask so this solution is
much improved.