Add CSV output
To make use of our x,y coordinate, we build a 2D array that can hold a
pretty big basement and add all of the string data we want to see in our
CSV to array at the appropriate coordinates.  A generic method to output
the array as a CSV spits the data into a file that looks nice in Excel
when we paste a generic format over it.