Add quality text and fix aging days bug
Figured out that the bug in the days to maturiy is caused by the fact that all items have a default days to
maturity of 56 (two seasons) but that items actually age towards this
value at different rates (wine=1, cheese=4, beer=1.66) etc.  Thus, the
proper value can be calculated by dividing by the aging rate.