Add test code for new Excel output with line break
By default, all data in a CSV is displayed in Excel cells on single
lines.  In order to make it actually do multi-line display properly, we
need to wrap the contents of the csv "cells" with quotes.  If we do
that, new lines in the text file are respected as new lines within the
Excel cells.