Refactor Cask loop to select only casks with linq
Going through every object is a waste if we only care about casks.  To
prepare for more modular code in the future, we can directly use linq to
select casks directly from doc.Descendants("Object") with .Where to
select only objects that have an xsi:type of Cask (or any other
arbitrary boolean expression in the future).  Much cleaner.