Refactor cask logic into a Cask class
Started modularizing the functionality in Stardew.cs by moving the
basement code into its' own method with the document as a parameter.  We
pass the collection of Cask save data objects into the Cask class in
order to construct a list of StardewTools Cask objects.

The cask class contains all of the logic necessary to parse data out of
the cask XElement in its' constructor, and currently contains the
methods needed to format the data in the way we previously did in the
main loop.

Other miscellaneous changes include offsetting the results of the
basement output so that the it's positioned with respect to cell A1 in
Excel.