Merge branch first-love into master (DogLove mod)
The DogLove mod was in its' own branch but that was squashed out of this
repository.  Future branches won't have this happen to them. The
modification is my first foray into SMAPI and simply displays in a log
message and HUD message if the player's pet was pet that day upon
pressing TAB.